+++ meau_app/App.tsx
@@ -1,5 +1,5 @@
 
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 
 import 'react-native-gesture-handler';
 
@@ -8,7 +8,7 @@ import 'react-native-gesture-handler';
 import * as SplashScreen from 'expo-splash-screen';
 
 import Routes from './src/routes/Routes';
-import { AutenticacaoUserProvider } from './assets/contexts/AutenticacaoUserContext';
+import { AutenticacaoUserProvider } from './src/assets/contexts/AutenticacaoUserContext';
 import { auth, getAuth, onAuthStateChanged} from './src/configs/firebaseConfig';
 
 
@@ -19,28 +19,13 @@ SplashScreen.preventAutoHideAsync()
 
 
 export default function App() {
-    const [ user, setUser ] = useState<any>(); //any para ajustar conforme o necessario
-    
-    useEffect(() => {
-        
-        const unsubscribe = onAuthStateChanged(getAuth(), (user) => {
-            console.log("Teste USUARIO 2"+ "estadoUser: " + user);
-            if (user) {
-                setUser(user); // Atualiza o estado com o usuário autenticado
-                console.log("Usuario logado: " + user.email);
-            } else {
-                setUser(null);  // Atualiza o estado para null se não houver usuário
-                console.log("Usuario off ");
-            }
-            SplashScreen.hideAsync();
-        });
-        return () => unsubscribe(); // Limpeza do listener ao desmontar o componente
-    }, []);
-    
 
+    useEffect(()=> {
+        SplashScreen.hideAsync();
+    }, []);
 
     return (
-        <AutenticacaoUserProvider value={{ user, setUser }}>
+        <AutenticacaoUserProvider>
             <Routes/>
         </AutenticacaoUserProvider>
     );
