+++ meau_app/src/routes/Routes.tsx
@@ -3,7 +3,7 @@ import { NavigationContainer } from "@react-navigation/native";
 import StackRoutes from "./stack/StackRoutes";
 import AuthStack from "./stack/AuthStack";
 
-import { useAutenticacaoUser } from "../../assets/contexts/AutenticacaoUserContext";
+import { useAutenticacaoUser } from "../assets/contexts/AutenticacaoUserContext";
 
 
 export default function Routes() {
